Organise imports using isort
Use `black`'s profile as that's pretty close to what we have and
would be forwards compatible with us picking either `black` or
`ruff` as a formatter in future.
