+++ mockafka/__init__.py
@@ -1,18 +1,18 @@
 from __future__ import annotations
 
 from .admin_client import FakeAdminClientImpl
-from .producer import FakeProducer
 from .consumer import FakeConsumer
-from .message import Message
 from .decorators import (
-    produce,
-    bulk_produce,
-    setup_kafka,
-    consume,
-    asetup_kafka,
     aconsume,
     aproduce,
+    asetup_kafka,
+    bulk_produce,
+    consume,
+    produce,
+    setup_kafka,
 )
+from .message import Message
+from .producer import FakeProducer
 
 __all__ = [
     "FakeProducer",
