+++ mockafka/cluster_metadata.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import logging
+
 from mockafka.broker_metadata import BrokerMetadata
 from mockafka.kafka_store import KafkaStore
 from mockafka.topic_metadata import TopicMetadata
