+++ mockafka/decorators/typing.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from typing import TypedDict
+
 from typing_extensions import NotRequired
