+++ mockafka/kafka_store.py
@@ -18,9 +18,11 @@ offset_store = {
 
 from __future__ import annotations
 
-from confluent_kafka import KafkaException  # type: ignore[import-untyped]
-from .message import Message
 from copy import deepcopy
+
+from confluent_kafka import KafkaException  # type: ignore[import-untyped]
+
+from .message import Message
 
 mock_topics: dict[str, dict[int, list[Message]]] = {}
