+++ pyproject.toml
@@ -21,6 +21,7 @@ typing-extensions = "^4.12.2"
 
 [tool.poetry.group.dev.dependencies]
 flake8 = "*"
+flake8-isort = "*"
 parameterized = "^0.9.0"
 pytest = ">=7.2,<9.0"
 pytest-cov = ">=4,<6"
@@ -36,6 +37,9 @@ mkdocstrings = { extras = ["python"], version = ">=0.23,<0.27" }
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.isort]
+profile = "black"
 
 [tool.mypy]
 warn_unused_ignores = true
