+++ tests/test_docrators.py
@@ -1,11 +1,12 @@
 from __future__ import annotations
 
-from mockafka import FakeConsumer, produce, bulk_produce, setup_kafka, Message
-from mockafka.admin_client import FakeAdminClientImpl, NewTopic
-from mockafka.producer import FakeProducer
-from mockafka.decorators.typing import MessageDict
-from mockafka.decorators.consumer import consume
 from unittest import TestCase
+
+from mockafka import FakeConsumer, Message, bulk_produce, produce, setup_kafka
+from mockafka.admin_client import FakeAdminClientImpl, NewTopic
+from mockafka.decorators.consumer import consume
+from mockafka.decorators.typing import MessageDict
+from mockafka.producer import FakeProducer
 
 sample_for_bulk_produce: list[MessageDict] = [
     {
